feat: support dynamic chains
