+++ .env.sample
@@ -25,18 +25,6 @@ MOCHI_BOT_SECRET='mochi_secret'
 
 JWT_SECRET='sample_secret'
 
-FANTOM_RPC='https://rpc.ftm.tools'
-FANTOM_SCAN='https://api.ftmscan.com/api?'
-FANTOM_SCAN_API_KEY='EAb28711eD07E0dB5F475748F88A664F58'
-
-ETHEREUM_RPC='https://mainnet.infura.io/v3/5b389eb75c514cf6b1711d70084b0114'
-ETHEREUM_SCAN='https://api.etherscan.io/api?'
-ETHEREUM_SCAN_API_KEY='SM5BHYSNIRZ1HEWJ1JPHVTMJS95HRA6DQF'
-
-BSC_RPC='https://bsc-dataseed.binance.org'
-BSC_SCAN='https://api.bscscan.com/api?'
-BSC_SCAN_API_KEY='VTKF4RG4HP6WXQ5QTAJ8MHDDIUFYD6VZHC'
-
 REDIS_URL="redis://localhost:6379/0"
 
 MOCHI_LOG_CHANNEL_ID="895659000996200511"
