+++ pkg/repo/chain/store.go
@@ -1,0 +1,7 @@
+package chain
+
+import "github.com/defipod/mochi/pkg/model"
+
+type Store interface {
+	GetAll() ([]model.Chain, error)
+}
