+++ pkg/repo/token/pg.go
@@ -15,20 +15,20 @@ func NewPG(db *gorm.DB) Store {
 
 func (pg *pg) GetBySymbol(symbol string, botSupported bool) (model.Token, error) {
 	var token model.Token
-	return token, pg.db.First(&token, "lower(symbol) = lower(?) AND discord_bot_supported = ?", symbol, botSupported).Error
+	return token, pg.db.Preload("Chain").First(&token, "lower(symbol) = lower(?) AND discord_bot_supported = ?", symbol, botSupported).Error
 }
 
 func (pg *pg) GetAllSupported() ([]model.Token, error) {
 	var tokens []model.Token
-	return tokens, pg.db.Where("discord_bot_supported = TRUE").Find(&tokens).Error
+	return tokens, pg.db.Preload("Chain").Where("discord_bot_supported = TRUE").Find(&tokens).Error
 }
 
 func (pg *pg) GetByAddress(address string, chainID int) (*model.Token, error) {
 	var token model.Token
-	return &token, pg.db.First(&token, "lower(address) = lower(?) and chain_id = ?", address, chainID).Error
+	return &token, pg.db.Preload("Chain").First(&token, "lower(address) = lower(?) and chain_id = ?", address, chainID).Error
 }
 
 func (pg *pg) GetDefaultTokens() ([]model.Token, error) {
 	var tokens []model.Token
-	return tokens, pg.db.Where("guild_default = TRUE").Find(&tokens).Error
+	return tokens, pg.db.Preload("Chain").Where("guild_default = TRUE").Find(&tokens).Error
 }
